requirements.txt
pandas~=1.4.3
nltk~=3.7
spacy~=3.3.1
#torch~=1.11.0
huggingface~=0.0.1
tensorboard~=2.9.1
transformers~=4.20.1
matplotlib~=3.5.2
seaborn~=0.11.2
scikit-learn~=1.1.1
gensim~=4.2.0
hyperopt~=0.2.7
tqdm~=4.64.0
joblib~=1.1.0
numpy~=1.23.0
tweet-preprocessor~=0.6.0
transformers~=4.20.1
torchtext~=0.12.0